+++ app/charts/column/columns-grouped-state-props.ts
@@ -7,6 +7,7 @@ import {
   BandXVariables,
   BaseVariables,
   ChartStateData,
+  InteractiveFiltersVariables,
   NumericalYErrorVariables,
   NumericalYVariables,
   RenderingVariables,
@@ -15,12 +16,14 @@ import {
   useBandXVariables,
   useBaseVariables,
   useChartData,
+  useInteractiveFiltersVariables,
   useNumericalYErrorVariables,
   useNumericalYVariables,
   useSegmentVariables,
 } from "@/charts/shared/chart-state";
 import { useRenderingKeyVariable } from "@/charts/shared/rendering-utils";
 import { ColumnConfig, useChartConfigFilters } from "@/configurator";
+import { Observation, isTemporalEntityDimension } from "@/domain/data";
 import { sortByIndex } from "@/utils/array";
 
 import { ChartProps } from "../shared/ChartProps";
@@ -31,7 +34,8 @@ export type ColumnsGroupedStateVariables = BaseVariables &
   NumericalYVariables &
   NumericalYErrorVariables &
   SegmentVariables &
-  RenderingVariables;
+  RenderingVariables &
+  InteractiveFiltersVariables;
 
 export const useColumnsGroupedStateVariables = (
   props: ChartProps<ColumnConfig>
@@ -46,6 +50,7 @@ export const useColumnsGroupedStateVariables = (
   } = props;
   const { fields, interactiveFiltersConfig } = chartConfig;
   const { x, y, segment, animation } = fields;
+  const xDimension = dimensionsByIri[x.componentIri];
   const filters = useChartConfigFilters(chartConfig);
 
   const baseVariables = useBaseVariables(chartConfig);
@@ -65,31 +70,38 @@ export const useColumnsGroupedStateVariables = (
     dimensionsByIri,
     observations,
   });
+  const interactiveFiltersVariables = useInteractiveFiltersVariables(
+    interactiveFiltersConfig,
+    { dimensionsByIri }
+  );
 
-  const { getX } = bandXVariables;
+  const { getX, getXAsDate } = bandXVariables;
   const { getY } = numericalYVariables;
   const sortData: ColumnsGroupedStateVariables["sortData"] = useCallback(
     (data) => {
       const { sortingOrder, sortingType } = x.sorting ?? {};
+      const xGetter = isTemporalEntityDimension(xDimension)
+        ? (d: Observation) => getXAsDate(d).getTime().toString()
+        : getX;
       const order = [
         ...rollup(
           data,
           (v) => sum(v, (d) => getY(d)),
-          (d) => getX(d)
+          (d) => xGetter(d)
         ),
       ]
         .sort((a, b) => ascending(a[1], b[1]))
         .map((d) => d[0]);
 
       if (sortingType === "byDimensionLabel") {
-        return orderBy(data, getX, sortingOrder);
+        return orderBy(data, xGetter, sortingOrder);
       } else if (sortingType === "byMeasure") {
-        return sortByIndex({ data, order, getCategory: getX, sortingOrder });
+        return sortByIndex({ data, order, getCategory: xGetter, sortingOrder });
       } else {
-        return orderBy(data, getX, "asc");
+        return orderBy(data, xGetter, "asc");
       }
     },
-    [getX, getY, x.sorting]
+    [getX, getXAsDate, getY, x.sorting, xDimension]
   );
 
   const getRenderingKey = useRenderingKeyVariable(
@@ -106,6 +118,7 @@ export const useColumnsGroupedStateVariables = (
     ...numericalYVariables,
     ...numericalYErrorVariables,
     ...segmentVariables,
+    ...interactiveFiltersVariables,
     getRenderingKey,
   };
 };
@@ -115,8 +128,13 @@ export const useColumnsGroupedStateData = (
   variables: ColumnsGroupedStateVariables
 ): ChartStateData => {
   const { chartConfig, observations } = chartProps;
-  const { sortData, getXAsDate, getY, getSegmentAbbreviationOrLabel } =
-    variables;
+  const {
+    sortData,
+    getXAsDate,
+    getY,
+    getSegmentAbbreviationOrLabel,
+    getTimeRangeDate,
+  } = variables;
   const plottableData = usePlottableData(observations, {
     getY,
   });
@@ -127,6 +145,7 @@ export const useColumnsGroupedStateData = (
     chartConfig,
     getXAsDate,
     getSegmentAbbreviationOrLabel,
+    getTimeRangeDate,
   });
 
   return {